+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Scripts/ExUnits.hs
@@ -3,7 +3,6 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -75,7 +74,7 @@ data TransactionScriptFailure era
       !ExUnits
       !P.EvaluationError
       ![Text]
-      !PlutusWithContext
+      !(PlutusWithContext (EraCrypto era))
   | -- | A redeemer points to a transaction input which is not
     --  present in the current UTxO.
     UnknownTxIn !(TxIn (EraCrypto era))
@@ -207,7 +206,7 @@ evalTxExUnitsWithLogs pp tx utxo epochInfo sysStart = do
         (\(sp, sh) -> (sp, lookupPlutusScript scriptsProvided sh, sh))
         scriptsNeeded
     findAndCount ptrToPlutusScript pointer (rdmr, exUnits) = do
-      (plutusPurpose, mPlutusScript, _) <-
+      (plutusPurpose, mPlutusScript, scriptHash) <-
         note (RedeemerPointsToUnknownScriptHash pointer) $
           Map.lookup pointer ptrToPlutusScript
       let ptrToPlutusScriptNoContext =
@@ -238,6 +237,7 @@ evalTxExUnitsWithLogs pp tx utxo epochInfo sysStart = do
               PlutusWithContext
                 { pwcProtocolVersion = protVerMajor
                 , pwcScript = Left plutus
+                , pwcScriptHash = scriptHash
                 , pwcDatums = PlutusDatums (getPlutusData <$> datums)
                 , pwcExUnits = maxBudget
                 , pwcCostModel = costModel